--- src/utils/checks.ts
@@ -1,81 +1,0 @@
-import { InvalidArgumentError } from 'commander'
-import fs from 'fs'
-import path from 'path'
-import { EnvMeta } from '../stubs/env-stubs'
-
-/**
- * Checks if the provided action path is valid
- *
- * @param value The action path
- * @returns The resolved action path
- */
-export function checkActionPath(value: string): string {
-  const actionPath: string = path.resolve(value)
-
-  try {
-    const stat: fs.Stats = fs.statSync(actionPath)
-
-    // Confirm the value is a directory
-    if (!stat.isDirectory())
-      throw new InvalidArgumentError('Action path must be a directory')
-  } catch (err: any) {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    if (err.code === 'ENOENT')
-      throw new InvalidArgumentError('Action path does not exist')
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-member-access
-    throw new InvalidArgumentError(err.message)
-  }
-
-  const actionFile: string = path.resolve(actionPath, 'action.yml')
-
-  // Confirm there is an `action.yml` in the directory
-  if (!fs.existsSync(actionFile))
-    throw new InvalidArgumentError(
-      'Action path must contain an action.yml file'
-    )
-
-  // Save the action path and file to environment metadata
-  EnvMeta.actionPath = actionPath
-  EnvMeta.actionFile = path.resolve(EnvMeta.actionPath, 'action.yml')
-
-  return path.resolve(value)
-}
-
-/**
- * Checks if the provided entrypoint is valid
- *
- * @param value The entrypoint
- * @returns The resolved entrypoint path
- */
-export function checkEntryPoint(value: string): string {
-  const entrypoint: string = path.resolve(EnvMeta.actionPath, value)
-
-  // Confirm the entrypoint exists
-  if (!fs.existsSync(entrypoint))
-    throw new InvalidArgumentError('Entrypoint does not exist')
-
-  // Save the action entrypoint to environment metadata
-  EnvMeta.entrypoint = entrypoint
-
-  return entrypoint
-}
-
-/**
- * Checks if the provided env file is valid
- *
- * @param value The env file path
- * @returns The resolved env file path
- */
-export function checkEnvFile(value: string): string {
-  const envFile: string = path.resolve(value)
-
-  // Confirm the env file exists
-  if (!fs.existsSync(envFile))
-    throw new InvalidArgumentError('Environment file does not exist')
-
-  // Save the .env file path to environment metadata
-  EnvMeta.envFile = envFile
-
-  return envFile
-}
